Change fuzzer UI to be in new information format
As we migrate to fuzzing more than just the SkPictures, this new data layout
should make it a bit more intuitive to use.  The basic idea is to slice
the information by the fuzzing category (e.g. skpicture, skcodec, api_foo, api_bar)
and then have the familiar "by file in the stacktrace" view.

fuzzer-summary-sk is the overview slice:
https://screenshot.googleplex.com/aYFgeNtppAK

Anything that collapses now uses the details-sk/summary-sk that jcgregorio ported.

fuzzer-summary-list-sk has been renamed to fuzzer-file-list-sk.  It only displays the
number of fuzzes and defaults to nothing expanded.  Note the URLs have been moved to
be on the file name, function name or line number.  Clicking on these will direct a user
to that specific slice (i.e. clicking on Line 566 will show only those fuzzes that are
in bar.h>frizzle()>Line566)
https://screenshot.googleplex.com/pgE4vskzsod

fuzzer-info-sk looks about the same, with the same url changes as before. It still defaults
to fully expanded.
https://screenshot.googleplex.com/aFEp2W3FR8s

Finally, the urls are to be changed about to be a bit more REST-like (changes will be
done server side on the next CL)
fuzzer.skia.org/category/:category/file/:file/func/:func/line/:line

Because of this change to the urls, everything but the categories will be Base-64 encoded
to ensure valid URLs.

BUG=skia:4438

Review URL: https://codereview.chromium.org/1638553002