Change fuzzer data model to be more easily sliced by type
This makes adding new categories to be fuzzed very easy (from the display side)
The Google storage layout changes to be
gs://[bucket]/:category/:skia-revision/:bad-or-grey/:fuzz-hash/:file

where files are the fuzz, the std errs, the dumps, and anything else we want to
add in.

Since the data has been disentangled, the fuzzcache needed to be updated to be
more granular with what was stored (one tree per category and a master list
of fuzz hashes for all categories at that revision).

Because the landing page (index.html) was changed to be the bad/grey counts,
fuzz_syncer was updated to fetch that information for each category.

BUG=skia:4438

Review URL: https://codereview.chromium.org/1636683002