debugger: Fix layout issues.
Make the center column acquire scrollbars if the image is too big to fit. Also add a vertical scrollbar to the Ops column so you can keep the image on the screen while viewing all the Ops.

Also moves the section to upload the SKP to the top, so it doesn't get pushed off the page.

Screenshot:

  https://screenshot.googleplex.com/wd8XcH0mkzP.png

BUG=skia:
TBR=ethannicholas

Review URL: https://codereview.chromium.org/1655903002