Port url-params to Polymer 1.0
Most of this was done with polyup.  I added the _loading to prevent the clobbering of the URL params when the page is loaded.  I am not sure if reflectToParams should actually be set, or if notify:true is sufficient.

I will be using this in the fuzzer.

BUG=skia:

Review URL: https://codereview.chromium.org/1664253002