Add select-status Polymer 1.x element
I will be using this in fuzzer, because iron-selector is too flakey
https://github.com/PolymerElements/iron-selector/issues/103
https://github.com/PolymerElements/iron-selector/issues/38

This element adds binding abilities to a select element and a clear button.  The clear button doesn't really work cleanly as its own element, so I lumped it in here.

The demo looks like:
https://screenshot.googleplex.com/ErnaK6DD3O7

BUG=skia:

Review URL: https://codereview.chromium.org/1666823003