Add AddressSanitizer to fuzzer analysis
Instead of running potential fuzzes against two executables (Debug, Release), this now runs them against 4 executables (Debug, Release) x (Clang, ASAN).  ASAN is short for a clang build with the AddressSanitizer flag on.

There are many new test cases that excercise the various combinations of ASAN output and Clang output.  For making a stacktrace, we look for the ASAN crash first, if that didn't crash, we use the Clang crash then. The ASAN stacktrace parsing requires 'llvm-symbolizer' to be on the path.  This typically involves a symbolic link.

To avoid duplicate code and functions with many parameters, I made some new structs in result.go, namely GCSPackage.  GCSPackage represents all the analysis information that is stored in Google Storage: The fuzz name and category, and the raw strings of the 6 files that are the output.  This allows reuse for uploading to and downloading from GCS.

A FuzzReport has changed from having a single HumanReadableFlags array to having one for each Debug and Release.  This allows for more symmetrical code and could allow for better filtering on the frontend.

There are two new flags for the backend, one that forces a re-analysis of the current fuzzes and one that toggles the outputs of builds.  The latter hopefully will make production logs easier to read.

BUG=skia:4438

Review URL: https://codereview.chromium.org/1668543004