Add UI to filter fuzzes based on tags.
There isn't a nice way to perform dynamic sums of child elements in polymer
(that I thought of, at least), so I resurrected fuzzer-collapse-function to make
the code easier to reason about.

This also bypasses the old "overview" which only showed the counts.  That
was very awkward with the filters, so I bypassed it.  Look for the deletion
in the next CL.

BUG=skia:4438

Review URL: https://codereview.chromium.org/1662373002