Filtering fuzzes by tag is now really really snappy.
The dom-ifs make this really snappy now, practically instant.  Filtering is now done in the trunk of the tree, not the leaves.

This includes a change to how select-status binds, as the old way didn't always work with actual Polymer elements.  Now any external changes to the values updates the selections properly.

BUG=skia:4438
BUG=skia:4753

Review URL: https://codereview.chromium.org/1680303003