Fuzzer now deduplicates on the analysis side instead of the download side
This will drastically decrease the stored fuzzes and reduce re-analysis time.

There were some refactorings in this, e.g. putting the logic to download all fuzzes in common/storage, moving frontend/data to data and putting the logic to download data.FuzzReports in a centralized location.

This also has a tiny UI bug that makes details-summary-sk notify when expanded is clicked.

Finally, a pipe "|" separated list of all fuzzes is uploaded to GCS when a commit is "closed", i.e. when we move from revision 'abc' to 'def', we store the names of the bad and grey fuzzes in 'abc' to speed up and make more accurate fuzz counting.
This will be built upon in a follow-on CL.

BUG=skia:4438

Review URL: https://codereview.chromium.org/1691893002