Store and retrieve ingestion meta data
This adds meta data storage for the ingestion process. Currently we keep track of which builds have been ingested. This information is then used to sync the ingested results with the trybot information from Rietveld and present a comprehensive view on the front-end about the state of the tryjobs.

https://screenshot.googleplex.com/ViPXTfU8D1p.png

BUG=skia:

Review URL: https://codereview.chromium.org/1676403003