Add api_parse_path fuzz
This also fixes a race condition in the aggregator, a caching problem in fuzz_syncer and allows for a different number of fuzz processes based on api vs binary.  It is likely the binary ones will have a larger search space and should have more processes.

BUG=skia:4438

Review URL: https://codereview.chromium.org/1706313002