Split up commits-sk into a few parts during transition to Polymer 1.x
The first two parts are commits-table and commits-data.  The former is only responsible for drawing the table and handling the UI components for filtering.  The latter fetches the data from the server and parses it.

Demo page:
https://screenshot.googleplex.com/c6s33u2tZuJ

The data layout/formats are the same as in the 0.5 version of Status.  The server shouldn't have to give out any different data than the 0.5 version.

Polymer 1.0 seems a bit faster than 0.5.  I was able to further increase the speed by dropping as many dom-ifs as I could.  From the docs "conditional templates are only useful to save initial creation cost when the elements being stamped are relatively heavyweight and the conditional may rarely (or never) be true in given usages. Otherwise, liberal use of conditional templates can actually add significant runtime performance overhead."

Time to load all builders for 16 commits with dom-ifs:
https://screenshot.googleplex.com/QkumVuEm9Cy
Time to load all builders for 16 commits with hidden$=:
https://screenshot.googleplex.com/MBpkEwcvcQp

Almost 2x is a good speed up.

BUG=skia:4254

Review URL: https://codereview.chromium.org/1742783002