Create buildbot-dash in Polymer 1.0
This adds an abstraction of the data fetching/filtering, a UI to choose the filters, a migration of input-list and wraps the entire thing up in app-sk.  There are some minor tweaks to build-table-sk, especially with how elements are deleted when a filter changes (good news, it's faster).

The main data format change was adding a way to uniquely identify builds and their respective "filler" builds [the divs that help align everything]

Full size:
https://screenshot.googleplex.com/DUDfP9wAX1z

Smaller screens:
https://screenshot.googleplex.com/gXgjcNkKS8p

BUG=skia:4254

Review URL: https://codereview.chromium.org/1773193002