Rewrite status commits-table using d3
The end result is something that is faster and (I think) easier to maintain.  D3 is much better suited for a data visualization like our Status page than Polymer is.  One major change is the drawing happens column by column instead of row by row.

While I was making changes, I added a mechanism to make it easier to see when builds stretch over branches.
https://screenshot.googleplex.com/KiwaYHQFbz3

Notice the dashed lines.

When all builds is selected, a scrollbar appears that scrolls the builds, but not the commits.
https://screenshot.googleplex.com/qHgt1c1oPta

The popups are now better aligned and the links work (darn camel case).

BUG=skia:4254
BUG=skia:5083

Review URL: https://codereview.chromium.org/1782473004