Do the actual compile and running of the fiddle sent from the browser.
Adds in the web assets to the push package.

Add some more detail to the DESIGN.md doc.

Add a hash cache to Builder.

Add in the rest of the security protections by locking down the nspawn container, and putting rlimits in place in fiddle_run.

BUG=skia:

Review URL: https://codereview.chromium.org/1889873004