Add Cloud logging scaffolding
This is meant for the RPIs to get logs to somewhere where we can see them, i.e. cloud logging.  This is the basic structure for scanning and parsing logs.  There are a few follow-on CLs which handle the actual pushing to GCL and deploying this to the PIs via pulld

BUG=skia:

Review-Url: https://codereview.chromium.org/2010493006