Fix datahopper "failed to delete metric" spew
We maintain a list of metrics we've created so that we can delete them,
but we never actually remove entries from that list after we delete
them. So the list just keeps growing forever, with the same entries
getting added over and over. The first time we see a metric in the list
we delete it successfully but subsequent deletes fail for obvious
reasons.

Also add more tests for MD5Params for warm'n'fuzzies.

TBR=rmistry
BUG=skia:5467

Review-Url: https://codereview.chromium.org/2104543002