Make hotspare actually reload image
This also changes hotspare to use rsync instead of scp because the former is usually more performant and efficient than the latter.

Additionally, this tones down the frequency of the syncs.  10 minutes should be fine.  1 minute was too aggressive.

I renamed hotspare/hotspare.go -> hotspare/main.go to be consistent with everything else in skolo (and most other places).

BUG=skia:

Review-Url: https://codereview.chromium.org/2141053002