Add count of exclusively covered digests to each ignore rule
This adds a column to the list of ignore rules that shows the number of digests exclusively covered by that rule (and no other rule). It has to be always smaller or the same as the already existing count of digests covered by the rule.

This is the first of two CL's. The next will add the ability to search for the digests that match this criterion.

BUG=skia:5613

Review-Url: https://codereview.chromium.org/2220573003