Add corpus selector to Gold
Makes the corpus indicators in the top bar selectable.
Corpus essentially becomes part of the context and is
automatically added to the query string where it makes sense.

Also fixes the jsonByBlameHandler to respect the source_type/corpus in the query parameter. Enabling by blame for corpora other than 'gm'.

Small changes to add consistency within the source.

BUG=skia:5642

Review-Url: https://codereview.chromium.org/2247533002