task_scheduler local_db: Add version and modification time to task value.
Modification time is ignored for now, but I expect that we will need it in the future to ensure cached Tasks haven't been updated in the DB (e.g. update status from Swarming concurrently with adding a comment from Status). I don't want to store modification time in Task because it will be very slow to GOB-decode the entire DB value just to check last modification time.

BUG=skia:5626

Review-Url: https://codereview.chromium.org/2253943008