Adds tile sampler and tile serialization for Gold
This adds a utility to create a sample of the current state of gold. The sample includes everything that represents
the current state of Gold. It takes about 15 seconds to create a full copy of the present state of Gold.

Further it adds a binary format for a Gold tile that is about 20% of the same tile when serialized to JSON.

The next CL will use use this to add tests to search, the ignore module and the new diff store.

BUG=skia:5340

Review-Url: https://codereview.chromium.org/2239683003