[task scheduler] Repeatedly find the optimal candidate for each spec
Instead of finding all candidates and sorting, subdivide by spec name
and find the best candidate, insert it into a cache wrapper with a fake
pretend-to-insert layer, then find the second best for that task spec,
and so on until all candidates for each task spec have been correctly
prioritized. This ensures that we get a correct sequence of task
candidates which may bisect one another.

The toy perftest takes 4-9 seconds on average to cycle.

NOTRY=true
BUG=skia:5626

Review-Url: https://codereview.chromium.org/2297863002