Add flag to tune caching in diffstore
Adds a flag that allows to set the number of GiB that should be used 
for caching in the diffstore. It's not an accurate number, but a way 
too limit or increase the memory used.

Also adds a failsafe to the handler that serves images. If an image
of a requested digest is not in the disk cache it will fetch it 
from Google storage.

BUG=skia:5340

Change-Id: If465cab73827bf06ea245c763236491b4ea89e88
Reviewed-on: https://skia-review.googlesource.com/2503
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Stephan Altmueller <stephana@google.com>