Add grid backend with new tested search implementation
This is the first of several CLs. 

- adds a rudimentary implementation of a new backend for the
gridview that allows to compare arbitrary digests within a test
and keep filter by matching parameter values.

- adds testing to search which has been absent from it. 

Some core features are not implemented (e.g. sorting). These are 
captured with TODOs and will be implemented in the next CLs. 

Next CLs will expose this as an endpoint, update the existing
fronte-end code and generalize the search function. 


BUG=skia:

Change-Id: I0f573c603535ca9d288cbb121208d2075c32c48a