Add gitsetup to testutils.
Use it for gitrepo test.

IMO, util.TempRepo makes tests very hard to understand because the files in the repo are hidden in an opaque zip file. This utility makes it easy to explicitly create files and commits in a repo so that it's obvious what data the test is working on.

BUG=skia:

Change-Id: Ieb23b604babdda39a5adb5281b27672e119791f3
Reviewed-on: https://skia-review.googlesource.com/2800
Commit-Queue: Ben Wagner <benjaminwagner@google.com>
Reviewed-by: Eric Boren <borenet@google.com>