Update patch set 1
Patch Set 1: Code-Review+1

I had assumed you had implemented the RE2 method so you could do matching in JavaScript.

Assuming a string-based match, I think the only thing that might improve on the current implementation is a prefix tree (although I would guess that's not as fast as a hash lookup, and not easy to implement efficiently in Go).

Patch-set: 1
Label: Code-Review=+1