perf: Make the query process async.
Mostly moves code for /_/frame/ out of main.go and breaks the 
single HTTP request into three.

Now that the code is broken out of main.go it can be tested,
which will come in a separate CL.

The only user visible change is the %-complete value next to the
spinner while requests are in process.

Screenshot:
  https://screenshot.googleplex.com/cky70etZZOH.png

BUG=skia:

Change-Id: I639bf787e1374b59d5292bc5467d2f08a5685810
Reviewed-on: https://skia-review.googlesource.com/3441
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Joe Gregorio <jcgregorio@google.com>