Update patch set 2
Patch Set 2:

(3 comments)

The <(VAR) syntax is actually borrowed from GYP.  No idea where they got it, if anywhere.  I'm open to other suggestions of how to handle this, but basically bot_update is not as tolerant of empty trybot-only args as gclient+apply_issue was, so we need a way to only add these args in that case.  I've been trying to be agnostic of these type of issues on the parsing side, but a really simple way would be to add an extra_args_trybot field and put the relevant stuff there.  Thoughts?

Patch-set: 2