Update patch set 1
Patch Set 1:

IMO manual tests should not exist. All tests should be run continuously, maybe not as frequently. What would be an example of a manual test ?

Having different (three is fine) levels of tests sgtm. Per our conversation, I assume the plan is to run on only the unit test on the CQ and run the other tests on the bots.

Patch-set: 1
Label: Code-Review=0