Add test types to testutils package
Add a call to UnitTest, IntegrationTest, or ManualTest as the first line
of yout test, and the test will be turned on or off by the --unittest,
--integration, and --manual flags. If none of those flags are specified,
UnitTests and IntegrationTests are run by default while ManualTests are
not. For example:

func TestMyFunc(t *testing.T) {
	testutils.UnitTest(t)

	assert.NoError(t, myfunc())
}

The test will run by default and when --unittest is specified. If any of
--unittest, --integration, or --manual is specified, then the defaults
no longer hold, and only the tests specified by the supplied flags will
run. For example, using --manual causes the UnitTests not to run, unless
--unittest is also specified.

This is totally opt-in and has no effect on existing tests. I chose not
to factor in test.Short for that reason. In the future I foresee running
each of the three sets of tests on a separate bot and having only
UnitTests as presubmit checks.

BUG=skia:

Change-Id: Icfb72449745fe33f445ae0825943668c4529bec6