Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (3 comments)
> 
> manual -> enormous in patch set 3.  Ravi, didn't you say that you had tests that were disabled because they took too long?  IMO that's the definition of an "enormous" test.  I have a test which actually should fall under the "manual" category: it's a big integration test which uploads a CL to code review to verify that we can patch in the CL (among other things).  We definitely don't want that running at every commit.

They take maybe about a minute to run but rely on different frameworks which is why they was disabled, "integration" (medium/large) would be perfect for it. "enormous" test (according to the definition) in the link times out at 1hr. "large" tests time out at 15mins.

Regarding your test to upload a CL to code review, where would that be turned on? if we do not want that running continuously then another option is to run such tests on a bot but only once a day. I also have similar tests that add comments to uploaded CLs. Maybe for us the definition of "enormous" tests should be end-to-end integration tests (like uploading CLs/modifying CLs/etc) that run once a day.

Patch-set: 3