Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > (3 comments)
> > 
> > manual -> enormous in patch set 3.  Ravi, didn't you say that you had tests that were disabled because they took too long?  IMO that's the definition of an "enormous" test.  I have a test which actually should fall under the "manual" category: it's a big integration test which uploads a CL to code review to verify that we can patch in the CL (among other things).  We definitely don't want that running at every commit.
> 
> They take maybe about a minute to run but rely on different frameworks which is why they was disabled, "integration" (medium/large) would be perfect for it. "enormous" test (according to the definition) in the link times out at 1hr. "large" tests time out at 15mins.
> 
> Regarding your test to upload a CL to code review, where would that be turned on? if we do not want that running continuously then another option is to run such tests on a bot but only once a day. I also have similar tests that add comments to uploaded CLs. Maybe for us the definition of "enormous" tests should be end-to-end integration tests (like uploading CLs/modifying CLs/etc) that run once a day.

Okay, I'm thinking that the test types should be based on both running time and types of dependencies. A unit test or small test is one with no external dependencies (except maybe filesystem access although that's disallowed in the linked doc).  An integration or medium test is one with some external dependencies OR which is too slow to be in the previous category.  The last category is kind of ambiguous but in my head it's "even slower" or "even more dependencies" than the middle category.  Maybe we just call them small/medium/large for simplicity and to avoid implying anything about what they actually do (ie. integration).  IMO the cutoffs are something like < 2 seconds for small tests, 2-15 seconds for medium tests, and > 15 seconds for large tests, but maybe it depends on how the existing tests get divided up.

As for my "manual" test, the problem is that I wouldn't want it running automatically at all, unless we had full support for cleaning up the uploaded CLs.  We'd want a non-production instance of the codereview server as well.  Frankly, the test isn't that important.  I ran it once to verify that it passed, and maybe I want to run it again if I modify that code in the future, but I'm fine with that being a manual process.  Anyway, I don't think we need to solve that problem for this first pass.  It's possible that I could rewrite the test to be saner, whether by somehow using a local codereview server or reusing the same CL over and over, etc.  In general I agree that manual-only tests should be avoided.

Patch-set: 3
Label: Code-Review=0