[CT] Clean tmp dir after each worker task.
This morning 90+ CT swarming bots were quarantined due to: "> 1024 files in TEMP (/tmp)"
Add a block that cleans tmp dir in all worker scripts.

All worker scripts used to have this block previously but it was removed during the conversion to swarming Eg:
https://codereview.chromium.org/2004153003/diff/260001/ct/go/worker_scripts/run_chromium_perf/main.go
It was probably removed because we assumed that swarming would take care
of all bot maintenance, but that does not seem to be the case for tmp dir.

BUG=skia:

Change-Id: I397ba5cf13ca07486cb6bcfc47392cd628d57ebe