Update patch set 1
Patch Set 1: Commit-Queue+1

I put a curly brace in the wrong place.  Note that in a perfect world where we didn't need the cache, we wouldn't run into this case, but all of the tasks for chrome/m55 branch have scrolled out of the cache time window, so we don't find them, resulting in our hitting the blamelist size limit. Maybe we should consider ensuring that all tasks associated with branch heads remain in the cache regardless of age?  Unfortunately I think that would require a linear scan of the database.

Patch-set: 1