Add test types to testutils package
Add a call to SmallTest, MediumTest, or LargeTest as the first line
of your test, and the test will be turned on or off by the --small,
--medium, and --large flags. If none of those flags are specified,
SmallTests and MediumTests are run by default while LargeTests are
not. For example:

func TestMyFunc(t *testing.T) {
	testutils.SmallTest(t)

	assert.NoError(t, myfunc())
}

The test will run by default and when --small is specified. If any of
--small, --medium, or --large is specified, then the defaults
no longer hold, and only the tests specified by the supplied flags will
run. For example, using --large causes the SmallTests not to run, unless
--small is also specified.

This is totally opt-in and has no effect on existing tests. I chose not
to factor in test.Short for that reason. In the future I foresee running
each of the three sets of tests on a separate bot and having only
SmallTests as presubmit checks.

BUG=skia:

Change-Id: Icfb72449745fe33f445ae0825943668c4529bec6