Fix disappearing comments on Status
Apparently there is some bogus data in the builder comments by builder
name table of the datahopper DB. As written, when searching for comments
for a particular builder, we start with the builder name as key prefix
but stop looking as soon as builder name + "|" no longer matches the
current key. I suspect that we somehow inserted data into that bucket
with a key of only the builder name, without the pipe, which causes us
to immediately exit the loop without picking up any of the comments.
Setting the loop to start at the correct prefix (builder + pipe) fixed
the issue.

BUG=skia:5941

Change-Id: Ie965ff69287971555439d89e3309c5fc52d9350b