Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

Code LGTM but what's the use case?  We can't run this on the production DB instance, so we'd have to download one of the backups.  Would it be terribly disruptive to run this within the task scheduler process, and maybe add a (googler-only) web front-end for inspecting the database?

Patch-set: 2
Reviewer: Eric Boren <5015@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Code-Review=+1