Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > FTR the most correct way to do this is to lock the TryJobIntegrator from performing sendHeartbeats for the entirety of TaskScheduler.updateUnfinishedJobs.
> 
> Isn't that what you're doing, since t.finishedMtx is locked for the duration of both of those methods?
> 
> Thus, I don't think you need t.finished.

No, t.finishedMtx is only locked for the duration of each JobFinished call; it's not locked during the cache update at the beginning of updateUnfinishedJobs, and it's not locked during the DB insert at the end.  I'm going to take the better approach, hang on a few minutes.

Patch-set: 1
Label: Code-Review=0