Update patch set 3
Patch Set 3:

I'm also not a huge fan of PS3.

I now understand PS1. The problem you solved with t.finished is the case where sendHeartbeats is called concurrently with updateUnfinishedJobs after the last call to jobFinished but before PutJobs. t.finished is like a "preview" of what's about to be given to PutJobs. This approach seems fine to me.

I would also be ok with two other approaches:
 - Add PutJobs method to JobCache that calls JobCache.Update at the end. Only call DB.PutJobs via JobCache.
 - Add BuildbucketStatus field to Job. Remove TryJobIntegrator.JobFinished and TryJobIntegrator.jobStarted, and instead use DB.GetModifiedJobs to notice when a Job's status doesn't match its BuildbucketStatus, and update Buildbucket. Maybe also check during sendHeartbeats. Eventual consistency FTW!

Patch-set: 3
Label: Code-Review=0