Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> not familiar with this code, but the change scares me a bit. Having a shared lock may cause a deadlocking code now or with future changes, and also will probably result in contention since lock is acquired for long periods of time (RPCs to external systems)

Agreed, it's not ideal.  We can avoid having a shared lock by running sendHeartbeats in the same goroutine as MainLoop, but I think that will dramatically slow MainLoop down, and I don't think this problem is bad enough to justify that.  We could move the locking by funneling writes through the cache or adding a CancelJob method to the DB, but that wouldn't solve this particular race, it'd only make us succeed in canceling the already-successful job.

Patch-set: 3