Don't do clustering on traces with <50% data.
Continuous clustering is up and running, but there are too many
false positives because for some bots (Win) there is insufficient
data. See:
  https://screenshot.googleplex.com/9j6Vr9xY8pD.png

For example, this alerting cluster is a set of traces with
only two points per trace:
  https://perf.skia.org/e/?keys=5889&begin=1479839155&end=1479847743

For now we don't cluster on traces with insufficient data. Going
forward we need to add formulas we can use to find out which
bots are not producing enough data so we can add more bots.

BUG=skia:

Change-Id: Id2db1db132f6d6a3d533962b92f69698a8443bea