Update patch set 1
Patch Set 1:

scheduleTasks is the slowest part at the moment, taking ~20 seconds on average out of the 30 second average loop.  5 seconds of that is querying for free Swarming bots.  This restructuring should make things a little faster. The next logical step is to isolate and trigger tasks in parallel.

Patch-set: 1