Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > What about the per-repo range issue?
> 
> My guess is that there are periods where 100 commits in the infra repo is ~1000 commits in the skia repo. Is TS able to handle that load?
> 
> (Either way, I prefer to submit as-is and fix in a followup CL.)

Based on the metrics I'm seeing, the expensive parts are when we interact with Swarming/Isolate and updating the repos.  The parts which scale with the number of commits/candidates amounts to less than 15% of the cycle time.  It only really gets expensive on startup when the TaskCfgCache is empty and we have to check out every commit in range.  So I think we could handle a lot more load but I don't really think we need or want to.

Patch-set: 2
Label: Commit-Queue=0