Update patch set 2
Patch Set 2:

I do not have background on what this CL is doing. But some information:
The CT builders currently run on the skiabot-ct-dm-* slaves on babysitters with large disks (to cache the many SKPs locally to save time, especially for the 1M bot).
The builders schedule swarming tasks on the Chromium pool (not on the CT pool). They run CPU tests on Chromium GCE instances and GPU tests on Chromium's bare-metal machines.
Eg:
* https://uberchromegw.corp.google.com/i/client.skia/builders/Perf-Ubuntu-GCC-Golo-GPU-GT610-x86_64-Release-CT_BENCH_1k_SKPs/builds/2179
* https://uberchromegw.corp.google.com/i/client.skia/builders/Perf-Ubuntu-GCC-GCE-CPU-AVX2-x86_64-Release-CT_BENCH_1k_SKPs/builds/2155

The CT pool is used exclusively by ct.skia.org

Patch-set: 2