Add tracking of failed downloads
This adds tracking of failed images to the diffstore. If an image
cannot be loaded form GS or decoded (because the .png is corrupted) 
it will recorded and we can analyze the root cause. 

This re-adds some code that from the old FileDiffStore and uses
the newly created boltutil package to simplify DB operations. 

BUG=skia:

Change-Id: I7322f50b3e6ec4fd589958c7baac83438743d28d