Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)

See https://github.com/google/skia-buildbot/blob/master/bash/release.sh#L88

If SYSTEMD_TIMER is set only the timer will be restarted. The idea was that timer would restart the main app. It's kind of a delayed trigger. 
With this setup if you updated just rpi-backup it would not be restarted.

Patch-set: 2