[task scheduler] Fix bugs in tests.
TestAddTasksFailure was relying on the DB not being updated when PutTasks returns an error, but I hadn't yet implemented that for inMemoryTaskDB. Implemented for both inMemoryTaskDB and inMemoryJobDB in this CL.

Also added assertions to DB tests that had been omitted due to this issue. This caught another bug in inMemory(Task|Job)DB where PutTask was not making a copy of the task, causing the DB to be updated when the caller later modified the task.

This issue resulted in TestUpdateUnfinishedTasks not really testing much, because the tasks already matched the versions in the DB. Possibly same problem in other tests.

BUG=skia:
NOTRY=true

Change-Id: I47f99379f9525e879fbe9bba3d545bfac05eeac2
Reviewed-on: https://skia-review.googlesource.com/5768
Commit-Queue: Ben Wagner <benjaminwagner@google.com>
Reviewed-by: Eric Boren <borenet@google.com>