Update patch set 2
Patch Set 2:

I verified that the Swarming service account is able to publish to this topic, and I built a little program locally which subscribes to the topic (pull, not push) and prints the messages.  I'd like to see this work for a little while, then we can switch over to using this as the primary source of Task data, and add a "load all" safety net goroutine to run every 10 minutes or so.

Patch-set: 2
Label: Code-Review=0