Pipe repo_base all the way through
This allows the different pages (infra, status) to link to the right code.

Additionally, this fixes a bug where "master" was not linkable in the
commits canvas.

This removes the now-obsolete "find all tasks by this bot" link, which no
longer works because we don't specify "id" as a task tag anymore.

BUG=skia:

Change-Id: Ic98f4242b62034cee667214b70ef355ce594a966