Add JSON handler for querying capacity statistics.
This returns the total time needed per commit, the number of tasks/commit
and the number of devices we have that matches a given dimension set.

This should be sufficient to create a capacity table similar to
https://docs.google.com/spreadsheets/d/1iGYTb-iYN1DnpJk0nMvW_RjLADXtNImuZpNBlei69Ko/edit
but dynamically.

Additionally this makes it easy to see our capacity on compile bots
(i.e. gpu:none), which is something the table did not have.

BUG=skia:6050

Change-Id: I6995adfa8e45cf3f9d8cf9ff6e55155bc62bdaab