Move (nearly) all glog calls to sklog and make Fuzzer use cloud logging
sklog will fall back to glog if sklog has not been initialized, so it
is backwards compatible.

The only glog calls that remain are in:
LogPanic() in go/common/common.go
go/sklog/*

to make sure problems with cloud logging and panics are logged to disk.

New clients simply need to add:
if !*local {
	c, err := auth.NewDefaultClient(false, sklog.CLOUD_LOGGING_WRITE_SCOPE)
	if err != nil {
		sklog.Fatalf("Problem getting authenticated client: %s", err)
	}
	err = sklog_init.InitCloudLogging(c, [app-name], [log-name])
	if err != nil {
		sklog.Fatalf("Could not setup cloud logging: %s", err)
	}
}

To their main(), just after the call to common.InitWithMetrics2.  Additionally,
the logging scope needs to be requested, which as of the time of this CL is
"https://www.googleapis.com/auth/logging.write"

Just like logserver (which will stick around as secondary logging), this puts
the number of lines of logs written into InfluxDB (e.g. for alerts).

BUG=skia:

Change-Id: Id994f80206d5247300551122cbad19707e43019c