Update patch set 3
Patch Set 3:

Regarding the circular import, maybe you could create a sklog.metricsClient interface so that the caller could pass in the metrics2.Client instance?

I also think that the InitCloudLogging should be part of common.Init.  I understand that it will take time to get all of the auth scopes set up, but that should be the eventual goal.

Patch-set: 3
Label: Code-Review=0