Update patch set 10
Patch Set 10:

(26 comments)

Thanks, this is awesome!

I can't give a +1 due to the infinite recursion issue (see below).

As long as we're also logging to disk, I can give a +1 even if you don't fix Flush/Fatal, but I think we need to fix it before moving to cloud-logging-only.

Patch-set: 10
Label: Code-Review=0