Fix alert tests for golang 1.7
In go 1.7, types.Eval seems to have had a slight behavior change:

Alert logic might be slightly broken in go 1.7+
https://play.golang.org/p/tB8H5hPv6a

From the docs:
https://golang.org/pkg/go/types/#Eval says "Eval should not be used 
instead of running Check to compute types and values, but in addition to Check. "

Basically, when trying to evaluate "x > a" when only "x" is defined,
go 1.6 would return an error on executing types.Eval() "undeclared name: a"

In go 1.7.3, types.Eval() no longer returns an error, but instead returns
a TypeAndValue with nil Type.

This updates the tests to handle both the 1.6 and the 1.7 logic until we
transfer to 1.7.

BUG=skia:

Change-Id: I9edbb88af08b56bec2c21b6f01f409b7fdc19441
Reviewed-on: https://skia-review.googlesource.com/6385
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Eric Boren <borenet@google.com>