Update patch set 1
Patch Set 1:

Just a thought from the peanut gallery... it'd be neat if somewhere in there we had "last rolled 7h ago".  It's sometimes just as relevant to the sheriff to know how stale the roll is as it is to know if it's failing or not.  I'm likely to caution the tree if the roll gets too stale, even if nothing is failing per se.

Patch-set: 1