[task scheduler] 3m trigger timeout, dismiss pubsub msg for dup tasks
I think what happened is that the task trigger endpoint took longer
than 1 minute to respond, so our HTTP client retried the request and the
second attempt succeeded. The first attempt eventually succeeded, so we
ended up with two Swarming tasks associated with the same task scheduler
task, which caused an error.

This CL extends the timeout to 3 minutes and changes the scheduler to
acknowledge the pubsub message and log an error, rather than returning
the error. That way we just ignore the extra task and get a single error
in the logs rather than a persistent stream of errors.

BUG=skia:6108

Change-Id: Ife1b4ddac5c636d584c49d83b217df7cdedd92d3