Remove hard coded references to Skia from gold front-end
The current use case for this pdfium-gold.skia.org.

This injects deployment specific variables into the 
index.html file that is used on all gold pages and 
removes hard coded references to the skia repository. 

Right now this only applies to the Git repo but will
probably be extended to also the Monorail and Gerrit URL. 

Other backend specific values can now easily be injected 
into the front-end templates by adding another field to the 
struct in main.go.

BUG=skia:5973

Change-Id: I7b3119fddc70292c795cc5d482c3fe8aeb649015