Update patch set 6
Patch Set 6:

> Patch Set 5:
> 
> Code change looks fine, but can we discuss the reason for needing the new fields?

I decided that it would be simpler not to store the isolate server and swarming server, since they will be the same for every task within an instance. If needed by Status, we could add them to Properties in Status based on the instance from which we get the data. URL is similar, with the added complication that we might need two different URLs depending on whether the user is logged-in or not.

Switched back to having a Fake() method rather than a Fake field.

Patch-set: 6